rename all `color` → `palette`
