+++ libs/@guardian/source/src/foundations/palette/palette.mdx
